Рефакторинг кода. Разделил утилитарный файл на несколько
