+++ src/mock/filter.js
@@ -1,4 +1,4 @@
-import {filter} from '../utils.js';
+import {filter} from '../utils/point-utils.js';
 
 export const generateFilter = (points) => Object.entries(filter).map(
   ([filterName, filterPoints]) => ({
